video/requirements.txt
python-dotenv
youtube-transcript-api
openai-whisper
groq
pytubefix
openai